Reformatted Tasks as a dialog
